Move class-validator tests to e2e tests directory
