+++ backend/src/modules/categories/categories.controller.spec.ts
@@ -105,70 +105,6 @@ describe('CategoriesController', () => {
       expect(result).toEqual(expected);
       expect(service.create).toHaveBeenCalledWith(createCategoryDto);
     });
-
-    it('should throw an error when passed unexpected props', async () => {
-      const createCategoryDto: CreateCategoryDto | any = {
-        name: 'Category Name',
-        notexpected: 'testnotexpected',
-      };
-
-      const categoryDtoObject = plainToInstance(
-        CreateCategoryDto,
-        createCategoryDto,
-      );
-
-      const errors: ValidationError[] = await validate(categoryDtoObject, {
-        whitelist: true,
-        forbidNonWhitelisted: true,
-      });
-
-      const expectedErrors: ValidationError[] = [
-        {
-          target: {
-            name: 'Category Name',
-            notexpected: 'testnotexpected',
-          },
-          value: 'testnotexpected',
-          property: 'notexpected',
-          children: undefined,
-          constraints: {
-            whitelistValidation: 'property notexpected should not exist',
-          },
-        },
-      ];
-
-      expect(errors).toEqual(expect.arrayContaining(expectedErrors));
-    });
-
-    it('should throw an error when passed invalid type props', async () => {
-      const createCategoryDto: CreateCategoryDto | any = {
-        name: 'Category Name',
-        parentId: true,
-      };
-
-      const categoryDtoObject = plainToInstance(
-        CreateCategoryDto,
-        createCategoryDto,
-      );
-
-      const errors: ValidationError[] = await validate(categoryDtoObject, {
-        whitelist: true,
-        forbidNonWhitelisted: true,
-        forbidUnknownValues: true,
-      });
-
-      const expectedErrors: ValidationError[] = [
-        {
-          target: { name: 'Category Name', parentId: true },
-          value: true,
-          property: 'parentId',
-          children: [],
-          constraints: { isUuid: 'parentId must be a UUID' },
-        },
-      ];
-
-      expect(errors).toEqual(expect.arrayContaining(expectedErrors));
-    });
   });
 
   describe('findAll', () => {
